moved components to components dir
