+++ components/page.jsx
@@ -1,0 +1,9 @@
+var React = require('react');
+
+var Page = React.createClass({
+  render: function() {
+    return <div className={this.props.name + "-page"}>{ this.props.children }</div>;
+  }
+});
+
+module.exports = Page;
